+++ docs/colors/modules/color-summary.svelte
@@ -14,7 +14,7 @@
 		padding: 6px 0;
 		width: 100%;
 		font-size: 12px;
-		color: var(--c-text-2);
+		color: var(--c-text-soft);
 		font-weight: 400;
 		text-transform: uppercase;
 		letter-spacing: 0.05em;