+++ sites/mytril.dev/src/lib/components/toolbar.svelte
@@ -103,7 +103,7 @@
 				font-size: 13px;
 				font-weight: 500;
 				text-decoration: inherit;
-				color: var(--c-brand);
+				color: var(--c-primary);
 				transition: color 0.5s;
 				line-height: 28px;
 				display: block;
@@ -125,7 +125,7 @@
 					font-weight: 500;
 					padding: 2px 0;
 					text-decoration: inherit;
-					color: var(--c-text-2);
+					color: var(--c-text-soft);
 					transition: color 0.5s;
 					line-height: 28px;
 					display: block;
@@ -134,7 +134,7 @@
 					white-space: nowrap;
 
 					&:hover {
-						color: var(--c-text-1);
+						color: var(--c-text);
 						transition: color 0.25s;
 					}
 				}